backend/api/routers/admin.py
"""
Admin Router
Administrative endpoints for system management.
"""

import logging
from typing import Dict

from fastapi import APIRouter
from pydantic import BaseModel

logger = logging.getLogger(__name__)

router = APIRouter()


class SystemStatus(BaseModel):
    """System status response"""

    status: str
    version: str
    components: Dict[str, str]


@router.get("/status", response_model=SystemStatus)
async def get_system_status():
    """
    Get system status.

    Returns system health and component status.
    """
    return SystemStatus(
        status="healthy",
        version="1.0.0",
        components={"api": "running", "database": "connected", "cache": "active"},
    )


@router.post("/maintenance")
async def enable_maintenance_mode():
    """
    Enable maintenance mode.

    Puts the system into maintenance mode.
    """
    logger.info("Maintenance mode enabled")
    return {
        "status": "maintenance_enabled",
        "message": "System is now in maintenance mode",
    }


@router.delete("/maintenance")
async def disable_maintenance_mode():
    """
    Disable maintenance mode.

    Takes the system out of maintenance mode.
    """
    logger.info("Maintenance mode disabled")
    return {"status": "maintenance_disabled", "message": "System is now operational"}


@router.get("/logs")
async def get_system_logs(limit: int = 100):
    """
    Get recent system logs.

    Args:
        limit: Maximum number of log entries to return

    Returns:
        Recent log entries
    """
    # Stub implementation
    return {"logs": [], "limit": limit, "message": "Log retrieval not yet implemented"}
