backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
requests
prometheus-client
redis
websockets
pydantic-settings
python-multipart
pyyaml>=6.0.1
watchdog>=3.0.0
plotly
aiohttp
river>=0.21.0,<0.22.0
polars>=1.0.0
connectorx>=0.3.0


docker==7.0.0
